+++ packages/activemodel/src/validations.trails.test.ts
@@ -1,0 +1,1881 @@
+import { describe, it, expect } from "vitest";
+import { Model } from "./index.js";
+import { resetI18n } from "./test-helpers/i18n.js";
+
+// Trails-only extras that have no `validations_test.rb` counterpart. They live
+// here rather than in `validations.test.ts` so the Rails-mirroring file stays a
+// one-to-one port of the Rails case.
+describe("ValidationsTest (trails)", () => {
+  // =========================================================================
+  // Phase 1100/1150 — Validations
+  // =========================================================================
+  // -- Presence --
+  describe("presence", () => {
+    class Article extends Model {
+      static {
+        this.attribute("title", "string");
+        this.validates("title", { presence: true });
+      }
+    }
+
+    it("rejects null", async () => {
+      const a = new Article();
+      expect(await a.isValid()).toBe(false);
+      expect(a.errors.get("title")).toContain("can't be blank");
+    });
+
+    it("rejects empty string", async () => {
+      const a = new Article({ title: "" });
+      expect(await a.isValid()).toBe(false);
+    });
+
+    it("rejects whitespace-only string", async () => {
+      const a = new Article({ title: "   " });
+      expect(await a.isValid()).toBe(false);
+    });
+
+    it("accepts a real value", async () => {
+      const a = new Article({ title: "Hello" });
+      expect(await a.isValid()).toBe(true);
+    });
+  });
+
+  // -- Absence --
+  describe("absence", () => {
+    class Blank extends Model {
+      static {
+        this.attribute("name", "string");
+        this.validates("name", { absence: true });
+      }
+    }
+
+    it("accepts null", async () => {
+      expect(await new Blank().isValid()).toBe(true);
+    });
+
+    it("accepts empty string", async () => {
+      expect(await new Blank({ name: "" }).isValid()).toBe(true);
+    });
+
+    it("rejects a value", async () => {
+      const b = new Blank({ name: "dean" });
+      expect(await b.isValid()).toBe(false);
+      expect(b.errors.get("name")).toContain("must be blank");
+    });
+  });
+
+  // -- Length --
+  describe("length", () => {
+    class WithLength extends Model {
+      static {
+        this.attribute("name", "string");
+        this.validates("name", {
+          length: { minimum: 3, maximum: 10 },
+        });
+      }
+    }
+
+    it("validates length of using minimum", async () => {
+      const w = new WithLength({ name: "ab" });
+      expect(await w.isValid()).toBe(false);
+      expect(w.errors.get("name")[0]).toMatch(/is too short/);
+    });
+
+    it("validates length of using maximum", async () => {
+      const w = new WithLength({ name: "abcdefghijk" });
+      expect(await w.isValid()).toBe(false);
+      expect(w.errors.get("name")[0]).toMatch(/is too long/);
+    });
+
+    it("validates length of using within", async () => {
+      expect(await new WithLength({ name: "dean" }).isValid()).toBe(true);
+    });
+
+    it("validates length of using is", async () => {
+      class Exact extends Model {
+        static {
+          this.attribute("code", "string");
+          this.validates("code", { length: { is: 4 } });
+        }
+      }
+      expect(await new Exact({ code: "1234" }).isValid()).toBe(true);
+      expect(await new Exact({ code: "123" }).isValid()).toBe(false);
+      expect(await new Exact({ code: "12345" }).isValid()).toBe(false);
+    });
+
+    it("validates with in (range)", async () => {
+      class WithRange extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { length: { in: [2, 5] } });
+        }
+      }
+      expect(await new WithRange({ name: "a" }).isValid()).toBe(false);
+      expect(await new WithRange({ name: "ab" }).isValid()).toBe(true);
+      expect(await new WithRange({ name: "abcde" }).isValid()).toBe(true);
+      expect(await new WithRange({ name: "abcdef" }).isValid()).toBe(false);
+    });
+
+    it("skips null values (null has no length)", async () => {
+      expect(await new WithLength({}).isValid()).toBe(false);
+    });
+  });
+
+  // -- Numericality --
+  describe("numericality", () => {
+    class Numeric extends Model {
+      static {
+        this.attribute("value", "string"); // string to test cast behavior
+        this.validates("value", { numericality: true });
+      }
+    }
+
+    it("default validates numericality of", async () => {
+      expect(await new Numeric({ value: "42" }).isValid()).toBe(true);
+      expect(await new Numeric({ value: "3.14" }).isValid()).toBe(true);
+    });
+
+    it("rejects non-numeric strings", async () => {
+      const n = new Numeric({ value: "not a number" });
+      expect(await n.isValid()).toBe(false);
+      expect(n.errors.get("value")).toContain("is not a number");
+    });
+
+    it("validates numericality of with nil allowed", async () => {
+      class NilOk extends Model {
+        static {
+          this.attribute("value", "string");
+          this.validates("value", { numericality: { allowNil: true } });
+        }
+      }
+      expect(await new NilOk({}).isValid()).toBe(true);
+    });
+
+    it("validates numericality of only integers", async () => {
+      class IntOnly extends Model {
+        static {
+          this.attribute("value", "string");
+          this.validates("value", { numericality: { onlyInteger: true } });
+        }
+      }
+      expect(await new IntOnly({ value: "42" }).isValid()).toBe(true);
+      expect(await new IntOnly({ value: "3.14" }).isValid()).toBe(false);
+    });
+
+    it("validates numericality with greater_than", async () => {
+      class GreaterThan extends Model {
+        static {
+          this.attribute("value", "integer");
+          this.validates("value", { numericality: { greaterThan: 5 } });
+        }
+      }
+      expect(await new GreaterThan({ value: 6 }).isValid()).toBe(true);
+      expect(await new GreaterThan({ value: 5 }).isValid()).toBe(false);
+    });
+
+    it("validates numericality with greater_than_or_equal_to", async () => {
+      class GreaterThanOrEqual extends Model {
+        static {
+          this.attribute("value", "integer");
+          this.validates("value", { numericality: { greaterThanOrEqualTo: 5 } });
+        }
+      }
+      expect(await new GreaterThanOrEqual({ value: 5 }).isValid()).toBe(true);
+      expect(await new GreaterThanOrEqual({ value: 4 }).isValid()).toBe(false);
+    });
+
+    it("validates numericality with equal_to", async () => {
+      class EqualTo extends Model {
+        static {
+          this.attribute("value", "integer");
+          this.validates("value", { numericality: { equalTo: 5 } });
+        }
+      }
+      expect(await new EqualTo({ value: 5 }).isValid()).toBe(true);
+      expect(await new EqualTo({ value: 4 }).isValid()).toBe(false);
+    });
+
+    it("validates numericality with less_than", async () => {
+      class LessThan extends Model {
+        static {
+          this.attribute("value", "integer");
+          this.validates("value", { numericality: { lessThan: 5 } });
+        }
+      }
+      expect(await new LessThan({ value: 4 }).isValid()).toBe(true);
+      expect(await new LessThan({ value: 5 }).isValid()).toBe(false);
+    });
+
+    it("validates numericality with less_than_or_equal_to", async () => {
+      class LessThanOrEqual extends Model {
+        static {
+          this.attribute("value", "integer");
+          this.validates("value", { numericality: { lessThanOrEqualTo: 5 } });
+        }
+      }
+      expect(await new LessThanOrEqual({ value: 5 }).isValid()).toBe(true);
+      expect(await new LessThanOrEqual({ value: 6 }).isValid()).toBe(false);
+    });
+
+    it("validates numericality with odd", async () => {
+      class Odd extends Model {
+        static {
+          this.attribute("value", "integer");
+          this.validates("value", { numericality: { odd: true } });
+        }
+      }
+      expect(await new Odd({ value: 3 }).isValid()).toBe(true);
+      expect(await new Odd({ value: 4 }).isValid()).toBe(false);
+    });
+
+    it("validates numericality with even", async () => {
+      class Even extends Model {
+        static {
+          this.attribute("value", "integer");
+          this.validates("value", { numericality: { even: true } });
+        }
+      }
+      expect(await new Even({ value: 4 }).isValid()).toBe(true);
+      expect(await new Even({ value: 3 }).isValid()).toBe(false);
+    });
+  });
+
+  // -- Inclusion / Exclusion --
+  describe("inclusion and exclusion", () => {
+    class Colorful extends Model {
+      static {
+        this.attribute("color", "string");
+        this.validates("color", {
+          inclusion: { in: ["red", "green", "blue"] },
+          exclusion: { in: ["black"] },
+        });
+      }
+    }
+
+    it("accepts included and non-excluded values", async () => {
+      expect(await new Colorful({ color: "red" }).isValid()).toBe(true);
+    });
+
+    it("rejects values not in inclusion list", async () => {
+      const c = new Colorful({ color: "yellow" });
+      expect(await c.isValid()).toBe(false);
+      expect(c.errors.get("color")).toContain("is not included in the list");
+    });
+
+    it("rejects values in exclusion list", async () => {
+      const c = new Colorful({ color: "black" });
+      expect(await c.isValid()).toBe(false);
+      expect(c.errors.get("color")).toContain("is reserved");
+    });
+  });
+
+  // -- Format --
+  describe("format", () => {
+    class EmailUser extends Model {
+      static {
+        this.attribute("email", "string");
+        this.validates("email", {
+          format: { with: /^[^@\s]+@[^@\s]+\.[^@\s]+$/, multiline: true },
+        });
+      }
+    }
+
+    it("accepts valid email", async () => {
+      expect(await new EmailUser({ email: "user@example.com" }).isValid()).toBe(true);
+    });
+
+    it("rejects invalid email", async () => {
+      const u = new EmailUser({ email: "invalid" });
+      expect(await u.isValid()).toBe(false);
+      expect(u.errors.get("email")).toContain("is invalid");
+    });
+  });
+
+  // -- Confirmation --
+  describe("confirmation", () => {
+    class Signup extends Model {
+      static {
+        this.attribute("password", "string");
+        this.attribute("passwordConfirmation", "string");
+        this.validates("password", { confirmation: true });
+      }
+    }
+
+    it("accepts matching password and confirmation", async () => {
+      expect(
+        await new Signup({ password: "secret", passwordConfirmation: "secret" }).isValid(),
+      ).toBe(true);
+    });
+
+    it("rejects mismatched password and confirmation", async () => {
+      const s = new Signup({ password: "secret", passwordConfirmation: "wrong" });
+      expect(await s.isValid()).toBe(false);
+      expect(s.errors.get("passwordConfirmation")).toContain("doesn't match Password");
+    });
+  });
+
+  // -- Uniqueness (simulated) --
+  describe("uniqueness", () => {
+    class UniqueUser extends Model {
+      static existingNames = new Set<string>();
+
+      static {
+        this.attribute("name", "string");
+        this.validates("name", {
+          presence: true,
+          uniqueness: true,
+        });
+      }
+
+      override async isValid(): Promise<boolean> {
+        const valid = await super.isValid();
+        if (!valid) return false;
+        const name = this.readAttribute("name") as string;
+        if (UniqueUser.existingNames.has(name)) {
+          // `errors.messages.taken` ships in activerecord's en.yml, not
+          // activemodel's — an Active Model-only case has to supply the literal.
+          this.errors.add("name", "has already been taken");
+          return false;
+        }
+        UniqueUser.existingNames.add(name);
+        return true;
+      }
+    }
+
+    it("accepts unique names", async () => {
+      UniqueUser.existingNames.clear();
+      expect(await new UniqueUser({ name: "alice" }).isValid()).toBe(true);
+      expect(await new UniqueUser({ name: "bob" }).isValid()).toBe(true);
+    });
+
+    it("rejects duplicate names", async () => {
+      UniqueUser.existingNames.clear();
+      const first = new UniqueUser({ name: "alice" });
+      const second = new UniqueUser({ name: "alice" });
+      expect(await first.isValid()).toBe(true);
+      expect(await second.isValid()).toBe(false);
+      expect(second.errors.get("name")).toContain("has already been taken");
+    });
+  });
+
+  // -- Type-based validations --
+  describe("type-based validations", () => {
+    class TypedModel extends Model {
+      static {
+        this.attribute("age", "integer");
+        this.attribute("email", "string");
+        this.validates("age", { presence: true, numericality: { onlyInteger: true } });
+        this.validates("email", { presence: true });
+      }
+    }
+
+    it("accepts valid types", async () => {
+      expect(await new TypedModel({ age: 30, email: "test@example.com" }).isValid()).toBe(true);
+    });
+
+    it("rejects invalid types", async () => {
+      const m = new TypedModel({ age: "not a number", email: "" } as any);
+      expect(await m.isValid()).toBe(false);
+      expect(m.errors.get("age").length).toBeGreaterThan(0);
+      expect(m.errors.get("email").length).toBeGreaterThan(0);
+    });
+  });
+
+  it("validates an undeclared getter via the send default", async () => {
+    // Rails' default `read_attribute_for_validation` is `send`, so a plain
+    // getter with no declared attribute is read by its accessor, not nil.
+    class Person extends Model {
+      static {
+        this.attribute("first", "string");
+      }
+      get fullName(): string {
+        return (this.readAttribute("first") as string) ?? "";
+      }
+    }
+    Person.validatesEach(["fullName"], (record, attr, value) => {
+      if (!value) record.errors.add(attr, "gotcha");
+    });
+    const present = new Person({ first: "Al" });
+    await present.isValid();
+    expect(present.errors.get("fullName")).not.toContain("gotcha");
+    const blank = new Person({ first: "" });
+    await blank.isValid();
+    expect(blank.errors.get("fullName")).toContain("gotcha");
+  });
+
+  it("read_attribute_for_validation returns undefined for a present reader that returns undefined", () => {
+    // Ruby `send` keys off method existence, not return value: a getter that
+    // exists and returns undefined yields undefined, it does not raise.
+    class Person extends Model {
+      static {
+        this.attribute("name", "string");
+      }
+      get optional(): string | undefined {
+        return undefined;
+      }
+    }
+    const p = new Person({ name: "Al" });
+    expect(
+      (
+        p as unknown as { readAttributeForValidation(a: string): unknown }
+      ).readAttributeForValidation("optional"),
+    ).toBeUndefined();
+  });
+
+  it("read_attribute_for_validation raises NoMethodError-style for a missing reader", () => {
+    // Ruby `send(:nope)` raises NoMethodError; a typo'd / undeclared validation
+    // attribute with no reader fails loud rather than validating a nil-ish value.
+    class Person extends Model {
+      static {
+        this.attribute("name", "string");
+      }
+    }
+    const p = new Person({ name: "Al" });
+    expect(() =>
+      (
+        p as unknown as { readAttributeForValidation(a: string): unknown }
+      ).readAttributeForValidation("nope"),
+    ).toThrow(/undefined method 'nope'/);
+  });
+
+  it("validates format of with multiline regexp should raise error", () => {
+    expect(() => {
+      class Person extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { format: { with: /^test$/m } });
+        }
+      }
+    }).toThrow(/multiline/i);
+  });
+
+  it("validates format of without any regexp should raise error", () => {
+    expect(() => {
+      class Person extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { format: {} });
+        }
+      }
+    }).toThrow(/with.*without/i);
+  });
+
+  describe("return-shape parity", () => {
+    class Topic extends Model {
+      static {
+        this.attribute("title", "string");
+        this.validates("title", { presence: true });
+      }
+    }
+
+    it("validate returns boolean (Rails alias_method :validate, :valid?)", async () => {
+      expect(await new Topic({ title: "ok" }).validate()).toBe(true);
+      expect(await new Topic({}).validate()).toBe(false);
+    });
+
+    it("invalid? accepts a context argument", async () => {
+      class Scoped extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true, on: "create" });
+        }
+      }
+      const s = new Scoped({});
+      expect(await s.isInvalid()).toBe(false);
+      expect(await s.isInvalid("create")).toBe(true);
+    });
+
+    it("validate! returns true and raises otherwise (never returns false)", async () => {
+      expect(await new Topic({ title: "ok" }).validateBang()).toBe(true);
+      await expect(new Topic({}).validateBang()).rejects.toThrow(/Validation failed/);
+    });
+
+    it("validate! forwards context to valid?", async () => {
+      // Rails validations.rb:417-419 — `valid?(context) || raise_validation_error`.
+      class Scoped extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true, on: "create" });
+        }
+      }
+      // No context → default context → no validators active → passes.
+      expect(await new Scoped({}).validateBang()).toBe(true);
+      // :create context → presence validator active → raises.
+      await expect(new Scoped({}).validateBang("create")).rejects.toThrow(/Validation failed/);
+    });
+
+    it("valid? accepts an array context that matches :on-registered validators", async () => {
+      // Rails `predicate_for_validation_context` (validations.rb:294-306)
+      // intersects the registered `on:` set with the model's current
+      // context — either side may be a single symbol or an array.
+      class Scoped extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("title", "string");
+          this.validates("name", { presence: true, on: "create" });
+          this.validates("title", { presence: true, on: ["publish"] });
+        }
+      }
+      // Single-symbol context → only the `on: :create` validator fires.
+      const a = new Scoped({});
+      expect(await a.isValid("create")).toBe(false);
+      expect(a.errors.attributeNames).toEqual(["name"]);
+
+      // Array context → both validators fire (Rails: intersection).
+      const b = new Scoped({});
+      expect(await b.isValid(["create", "publish"])).toBe(false);
+      expect(b.errors.attributeNames.sort()).toEqual(["name", "title"]);
+
+      // Array context matching only one registered value fires that one.
+      const c = new Scoped({});
+      expect(await c.isValid(["publish"])).toBe(false);
+      expect(c.errors.attributeNames).toEqual(["title"]);
+    });
+
+    it("on: [array] validator fires when current context is a single symbol in the set", async () => {
+      class Scoped extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true, on: ["create", "publish"] });
+        }
+      }
+      expect(await new Scoped({}).isValid("create")).toBe(false);
+      expect(await new Scoped({}).isValid("publish")).toBe(false);
+      expect(await new Scoped({}).isValid("unrelated")).toBe(true);
+    });
+
+    it("validationContext round-trips array contexts while a validation is in flight", async () => {
+      // Rails `validation_context` surfaces whatever context is currently
+      // set — when called inside a validator with an array context, it
+      // returns the array.
+      const captured: Array<string | string[] | null> = [];
+      class Scoped extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validate((record: InstanceType<typeof Scoped>) => {
+            captured.push(record.validationContext);
+          });
+        }
+      }
+      await new Scoped({}).isValid(["create", "publish"]);
+      expect(captured).toEqual([["create", "publish"]]);
+    });
+
+    it("valid?(null) clears the context (Rails sets it to nil on entry)", async () => {
+      // Rails `valid?(context = nil)` always assigns
+      // `context_for_validation.context = context` — passing nil clears.
+      const captured: Array<string | string[] | null> = [];
+      class Scoped extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validate((r: InstanceType<typeof Scoped>) => {
+            captured.push(r.validationContext);
+          });
+        }
+      }
+      const m = new Scoped({});
+      await m.isValid("previous");
+      await m.isValid(null);
+      // First call saw "previous"; second call saw null (explicit clear),
+      // not "previous" carried over.
+      expect(captured).toEqual(["previous", null]);
+    });
+
+    it("valid? restores previous context in ensure/finally even on failure", async () => {
+      // Rails validations.rb:361-368 uses `ensure` to restore context.
+      class Scoped extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      const m = new Scoped({});
+      const before = m.validationContext;
+      await m.isValid("custom");
+      expect(m.validationContext).toBe(before);
+    });
+  });
+
+  describe("ValidationError + freeze (Rails fidelity)", () => {
+    class Topic extends Model {
+      static {
+        this.attribute("title", "string");
+        this.validates("title", { presence: true });
+      }
+    }
+
+    it("ValidationError message comes from I18n :model_invalid", async () => {
+      // Default en → "Validation failed: %{errors}"
+      // (activemodel locale/en.yml:9).
+      await expect(new Topic({}).validateBang()).rejects.toThrow(
+        /^Validation failed: Title can't be blank$/,
+      );
+    });
+
+    it("ValidationError message picks up per-scope override", async () => {
+      const { I18n } = await import("./i18n.js");
+      I18n.backend().storeTranslations("en", {
+        activemodel: {
+          errors: {
+            messages: { model_invalid: "Nope: %{errors}" },
+          },
+        },
+      });
+      try {
+        await expect(new Topic({}).validateBang()).rejects.toThrow(/^Nope: /);
+      } finally {
+        resetI18n();
+      }
+    });
+
+    it("freeze locks the object and returns self", () => {
+      const t = new Topic({ title: "ok" });
+      expect(t.freeze()).toBe(t);
+      expect(Object.isFrozen(t)).toBe(true);
+    });
+
+    it("freeze preserves errors/validationContext access (Rails pre-touch)", () => {
+      const t = new Topic({ title: "ok" });
+      t.freeze();
+      // Rails validations.rb:372-377 ensures these lazy ivars are
+      // materialized so frozen models can still answer.
+      expect(t.errors).toBeDefined();
+      expect(t.validationContext).toBe(null);
+    });
+
+    it("contextForValidation is a live view of validationContext", () => {
+      const t = new Topic({ title: "ok" });
+      const vc = t.contextForValidation();
+      vc.context = "create";
+      expect(t.validationContext).toBe("create");
+      // Same instance is returned on subsequent calls (Rails ||= memoization).
+      expect(t.contextForValidation()).toBe(vc);
+    });
+
+    it("contextForValidation is callable on a frozen model", () => {
+      const t = new Topic({ title: "ok" });
+      t.freeze();
+      // freeze() pre-materializes the cache; access must not throw.
+      expect(() => t.contextForValidation()).not.toThrow();
+    });
+  });
+
+  describe("_validators hash-of-arrays (Rails fidelity)", () => {
+    // Rails `_validators = Hash.new { |h, k| h[k] = [] }`
+    // (activemodel/lib/active_model/validations.rb:50) — per-attribute
+    // buckets, O(1) `validators_on`, dup-in-inherited.
+    it("validatorsOn is O(1) per-attribute lookup", () => {
+      class Person extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("age", "integer");
+          this.validates("name", { presence: true });
+          this.validates("age", { numericality: true });
+        }
+      }
+      expect(Person.validatorsOn("name")).toHaveLength(1);
+      expect(Person.validatorsOn("age")).toHaveLength(1);
+      expect(Person.validatorsOn("nonexistent")).toEqual([]);
+    });
+
+    it("validators() returns a uniq flat list across all attribute buckets", () => {
+      class Person extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("email", "string");
+          // validates_each binds one validator across two attributes —
+          // it lands in both buckets and must still appear once via
+          // `validators()` (Rails: `_validators.values.flatten.uniq`).
+          this.validatesEach(["name", "email"], () => {});
+        }
+      }
+      expect(Person.validators()).toHaveLength(1);
+      expect(Person.validatorsOn("name")).toHaveLength(1);
+      expect(Person.validatorsOn("email")).toHaveLength(1);
+      expect(Person.validatorsOn("name")[0]).toBe(Person.validatorsOn("email")[0]);
+    });
+
+    it("inheritance is copy-on-first-write (subclass sees parent writes made before its own first write)", () => {
+      // Documented divergence from Rails. Rails' `inherited(base)` hook runs
+      // eagerly at `class Child < Base; end` time and snapshots
+      // `_validators`, so subsequent `Base.validates` additions don't reach
+      // `Child`. JS has no `inherited` hook that fires at subclass
+      // definition, so we defer the dup until Child's first write. In this
+      // window, Child still reads Base's Map via the prototype chain.
+      class Base extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      class Child extends Base {}
+      expect(Child.validatorsOn("name")).toHaveLength(1);
+      // Parent adds another validator AFTER Child is defined, and Child has
+      // not yet registered anything of its own. Copy-on-first-write
+      // semantics: Child still sees Base's map, so the new validator
+      // propagates.
+      Base.validates("name", { length: { minimum: 2 } });
+      expect(Child.validatorsOn("name")).toHaveLength(2);
+      // As soon as Child writes, it detaches. Further Base writes stay on
+      // Base.
+      Child.validates("name", { length: { maximum: 10 } });
+      Base.validates("name", { format: { with: /x/ } });
+      expect(Child.validatorsOn("name")).toHaveLength(3);
+      expect(Base.validatorsOn("name")).toHaveLength(3);
+      expect(Child.validatorsOn("name")).not.toContain(Base.validatorsOn("name")[2]);
+    });
+
+    it("subclass inherits validators but its changes don't leak up", () => {
+      class Base extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      class Child extends Base {}
+      // Subclass sees parent's validators…
+      expect(Child.validatorsOn("name")).toHaveLength(1);
+      // …and adding one on the subclass must not affect the parent.
+      Child.validates("name", { length: { minimum: 2 } });
+      expect(Child.validatorsOn("name")).toHaveLength(2);
+      expect(Base.validatorsOn("name")).toHaveLength(1);
+    });
+
+    it("clearValidators! empties the map", () => {
+      class Person extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      expect(Person.validators()).toHaveLength(1);
+      Person.clearValidatorsBang();
+      expect(Person.validators()).toEqual([]);
+      expect(Person.validatorsOn("name")).toEqual([]);
+    });
+
+    it("routes arbitrary { validate() } class into the right bucket via explicit attributes", () => {
+      // Rails `validates_with` also accepts any class that just implements
+      // `validate(record)`. Such a class won't expose `attributes` on the
+      // instance or in `options`, so `validatesWith` must route the explicit
+      // `attributes:` option through to the bucket lookup.
+      class PojoValidator {
+        validate(_record: unknown): void {
+          /* no-op */
+        }
+      }
+      class Person extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validatesWith(PojoValidator, { attributes: ["name"] });
+        }
+      }
+      expect(Person.validatorsOn("name")).toHaveLength(1);
+      expect(Person.validatorsOn("name")[0]).toBeInstanceOf(PojoValidator);
+    });
+
+    it("routes plain Validator with attributes: option into the right bucket", async () => {
+      // Rails `validates_with MyValidator, attributes: [:name]` — the
+      // validator is a plain `Validator` (not EachValidator); attributes
+      // live in `options` rather than directly on the instance.
+      // _registerValidator must check both.
+      const { Validator: ValidatorBase } = await import("./validator.js");
+      class StaticValidator extends ValidatorBase {
+        override validate(): void {
+          /* no-op */
+        }
+      }
+      class Person extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validatesWith(StaticValidator, { attributes: ["name"] });
+        }
+      }
+      expect(Person.validatorsOn("name")).toHaveLength(1);
+      expect(Person.validatorsOn("name")[0]).toBeInstanceOf(StaticValidator);
+    });
+
+    it("validatorsOn returns a fresh array (no state-mutating reads)", () => {
+      class Person extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      // Reading an unseen attribute must NOT create a bucket (unlike Rails'
+      // default-proc hash) — the TS API keeps reads side-effect-free.
+      Person.validatorsOn("never_registered");
+      expect(Array.from(Person._validators.keys())).not.toContain("never_registered");
+
+      // Mutating the returned array must NOT affect internal state.
+      const a = Person.validatorsOn("name");
+      a.length = 0;
+      expect(Person.validatorsOn("name")).toHaveLength(1);
+
+      // Consecutive calls return independent arrays.
+      expect(Person.validatorsOn("name")).not.toBe(Person.validatorsOn("name"));
+    });
+  });
+
+  describe("validatesEach", () => {
+    it("validates each", async () => {
+      class Payment extends Model {
+        static {
+          this.attribute("price", "integer");
+          this.attribute("discount", "integer");
+          this.validatesEach(["price", "discount"], (record, attr, value) => {
+            if (typeof value === "number" && value < 0) {
+              record.errors.add(attr, ":invalid", { message: "must be non-negative" });
+            }
+          });
+        }
+      }
+
+      const p = new Payment({ price: -5, discount: 10 });
+      expect(await p.isValid()).toBe(false);
+      expect(p.errors.fullMessages).toContain("Price must be non-negative");
+
+      const p2 = new Payment({ price: 5, discount: -3 });
+      expect(await p2.isValid()).toBe(false);
+      expect(p2.errors.fullMessages).toContain("Discount must be non-negative");
+
+      const p3 = new Payment({ price: 5, discount: 10 });
+      expect(await p3.isValid()).toBe(true);
+    });
+
+    it("supports conditional options", async () => {
+      class Payment extends Model {
+        static {
+          this.attribute("price", "integer");
+          this.attribute("discount", "integer");
+          this.validatesEach(
+            ["price", "discount"],
+            (record, attr, value) => {
+              if (typeof value === "number" && value < 0) {
+                record.errors.add(attr, ":invalid", { message: "must be non-negative" });
+              }
+            },
+            {
+              if: (record) =>
+                (record as unknown as { readAttribute(a: string): unknown }).readAttribute(
+                  "price",
+                ) !== null,
+            },
+          );
+        }
+      }
+
+      const p = new Payment({ price: null, discount: -3 });
+      expect(await p.isValid()).toBe(true);
+    });
+  });
+
+  describe("validatesWith", () => {
+    it("validation with class that adds errors", async () => {
+      class EvenValidator {
+        validate(record: any) {
+          const val = record.readAttribute("count");
+          if (typeof val === "number" && val % 2 !== 0) {
+            record.errors.add("count", ":invalid", { message: "must be even" });
+          }
+        }
+      }
+
+      class Item extends Model {
+        static {
+          this.attribute("count", "integer");
+          this.validatesWith(EvenValidator);
+        }
+      }
+
+      const item = new Item({ count: 3 });
+      expect(await item.isValid()).toBe(false);
+      expect(item.errors.fullMessages).toContain("Count must be even");
+
+      const item2 = new Item({ count: 4 });
+      expect(await item2.isValid()).toBe(true);
+    });
+
+    it("passes all configuration options to the validator class", async () => {
+      class ThresholdValidator {
+        threshold: number;
+        constructor(options: any = {}) {
+          this.threshold = options.threshold ?? 10;
+        }
+        validate(record: any) {
+          const val = record.readAttribute("score");
+          if (typeof val === "number" && val < this.threshold) {
+            record.errors.add("score", ":invalid", {
+              message: `must be at least ${this.threshold}`,
+            });
+          }
+        }
+      }
+
+      class Game extends Model {
+        static {
+          this.attribute("score", "integer");
+          this.validatesWith(ThresholdValidator, { threshold: 50 });
+        }
+      }
+
+      const g = new Game({ score: 30 });
+      expect(await g.isValid()).toBe(false);
+      expect(g.errors.fullMessages).toContain("Score must be at least 50");
+
+      const g2 = new Game({ score: 60 });
+      expect(await g2.isValid()).toBe(true);
+    });
+
+    it("supports conditional options", async () => {
+      class AlwaysInvalidValidator {
+        validate(record: any) {
+          record.errors.add("base", ":invalid", { message: "always invalid" });
+        }
+      }
+
+      class Widget extends Model {
+        static {
+          this.attribute("active", "boolean");
+          this.validatesWith(AlwaysInvalidValidator, {
+            if: (r: any) => r.readAttribute("active") === true,
+          });
+        }
+      }
+
+      const w = new Widget({ active: false });
+      expect(await w.isValid()).toBe(true);
+
+      const w2 = new Widget({ active: true });
+      expect(await w2.isValid()).toBe(false);
+    });
+  });
+  describe("Validations", () => {
+    // -- Presence --
+    describe("presence", () => {
+      class Article extends Model {
+        static {
+          this.attribute("title", "string");
+          this.validates("title", { presence: true });
+        }
+      }
+
+      it("rejects null", async () => {
+        const a = new Article();
+        expect(await a.isValid()).toBe(false);
+        expect(a.errors.get("title")).toContain("can't be blank");
+      });
+
+      it("rejects empty string", async () => {
+        const a = new Article({ title: "" });
+        expect(await a.isValid()).toBe(false);
+      });
+
+      it("rejects whitespace-only string", async () => {
+        const a = new Article({ title: "   " });
+        expect(await a.isValid()).toBe(false);
+      });
+
+      it("accepts a real value", async () => {
+        const a = new Article({ title: "Hello" });
+        expect(await a.isValid()).toBe(true);
+      });
+    });
+
+    // -- Absence --
+    describe("absence", () => {
+      class Blank extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { absence: true });
+        }
+      }
+
+      it("accepts null", async () => {
+        expect(await new Blank().isValid()).toBe(true);
+      });
+
+      it("accepts empty string", async () => {
+        expect(await new Blank({ name: "" }).isValid()).toBe(true);
+      });
+
+      it("rejects a value", async () => {
+        const b = new Blank({ name: "dean" });
+        expect(await b.isValid()).toBe(false);
+        expect(b.errors.get("name")).toContain("must be blank");
+      });
+    });
+
+    // -- Length --
+    describe("length", () => {
+      class WithLength extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", {
+            length: { minimum: 3, maximum: 10 },
+          });
+        }
+      }
+
+      it("validates length of using minimum", async () => {
+        const w = new WithLength({ name: "ab" });
+        expect(await w.isValid()).toBe(false);
+        expect(w.errors.get("name")[0]).toMatch(/is too short/);
+      });
+
+      it("validates length of using maximum", async () => {
+        const w = new WithLength({ name: "abcdefghijk" });
+        expect(await w.isValid()).toBe(false);
+        expect(w.errors.get("name")[0]).toMatch(/is too long/);
+      });
+
+      it("validates length of using within", async () => {
+        expect(await new WithLength({ name: "dean" }).isValid()).toBe(true);
+      });
+
+      it("validates length of using is", async () => {
+        class Exact extends Model {
+          static {
+            this.attribute("code", "string");
+            this.validates("code", { length: { is: 4 } });
+          }
+        }
+        expect(await new Exact({ code: "1234" }).isValid()).toBe(true);
+        expect(await new Exact({ code: "123" }).isValid()).toBe(false);
+        expect(await new Exact({ code: "12345" }).isValid()).toBe(false);
+      });
+
+      it("validates with in (range)", async () => {
+        class WithRange extends Model {
+          static {
+            this.attribute("name", "string");
+            this.validates("name", { length: { in: [2, 5] } });
+          }
+        }
+        expect(await new WithRange({ name: "a" }).isValid()).toBe(false);
+        expect(await new WithRange({ name: "ab" }).isValid()).toBe(true);
+        expect(await new WithRange({ name: "abcde" }).isValid()).toBe(true);
+        expect(await new WithRange({ name: "abcdef" }).isValid()).toBe(false);
+      });
+
+      it("skips null values (null has no length)", async () => {
+        expect(await new WithLength({}).isValid()).toBe(false);
+      });
+    });
+
+    // -- Numericality --
+    describe("numericality", () => {
+      class Numeric extends Model {
+        static {
+          this.attribute("value", "string"); // string to test cast behavior
+          this.validates("value", { numericality: true });
+        }
+      }
+
+      it("default validates numericality of", async () => {
+        expect(await new Numeric({ value: "42" }).isValid()).toBe(true);
+        expect(await new Numeric({ value: "3.14" }).isValid()).toBe(true);
+      });
+
+      it("rejects non-numeric strings", async () => {
+        const n = new Numeric({ value: "not a number" });
+        expect(await n.isValid()).toBe(false);
+        expect(n.errors.get("value")).toContain("is not a number");
+      });
+
+      it("validates numericality of with nil allowed", async () => {
+        class NilOk extends Model {
+          static {
+            this.attribute("count", "string");
+            this.validates("count", { numericality: { allowNil: true } });
+          }
+        }
+        expect(await new NilOk({}).isValid()).toBe(true);
+      });
+
+      it("validates numericality of with integer only", async () => {
+        class IntOnly extends Model {
+          static {
+            this.attribute("count", "string");
+            this.validates("count", { numericality: { onlyInteger: true } });
+          }
+        }
+        expect(await new IntOnly({ count: "5" }).isValid()).toBe(true);
+        const f = new IntOnly({ count: "5.5" });
+        expect(await f.isValid()).toBe(false);
+        expect(f.errors.get("count")).toContain("must be an integer");
+      });
+
+      it("validates numericality with greater than", async () => {
+        class GT extends Model {
+          static {
+            this.attribute("age", "integer");
+            this.validates("age", { numericality: { greaterThan: 0 } });
+          }
+        }
+        expect(await new GT({ age: 1 }).isValid()).toBe(true);
+        expect(await new GT({ age: 0 }).isValid()).toBe(false);
+      });
+
+      it("validates numericality with less than", async () => {
+        class LT extends Model {
+          static {
+            this.attribute("rating", "integer");
+            this.validates("rating", { numericality: { lessThan: 10 } });
+          }
+        }
+        expect(await new LT({ rating: 9 }).isValid()).toBe(true);
+        expect(await new LT({ rating: 10 }).isValid()).toBe(false);
+      });
+
+      it("validates numericality with odd", async () => {
+        class Odd extends Model {
+          static {
+            this.attribute("n", "integer");
+            this.validates("n", { numericality: { odd: true } });
+          }
+        }
+        expect(await new Odd({ n: 3 }).isValid()).toBe(true);
+        expect(await new Odd({ n: 4 }).isValid()).toBe(false);
+      });
+
+      it("validates numericality with even", async () => {
+        class Even extends Model {
+          static {
+            this.attribute("n", "integer");
+            this.validates("n", { numericality: { even: true } });
+          }
+        }
+        expect(await new Even({ n: 4 }).isValid()).toBe(true);
+        expect(await new Even({ n: 3 }).isValid()).toBe(false);
+      });
+    });
+
+    // -- Inclusion / Exclusion --
+    describe("inclusion", () => {
+      class Status extends Model {
+        static {
+          this.attribute("status", "string");
+          this.validates("status", { inclusion: { in: ["draft", "published"] } });
+        }
+      }
+
+      it("validates inclusion of", async () => {
+        expect(await new Status({ status: "draft" }).isValid()).toBe(true);
+      });
+
+      it("rejects non-included values", async () => {
+        const s = new Status({ status: "invalid" });
+        expect(await s.isValid()).toBe(false);
+        expect(s.errors.get("status")).toContain("is not included in the list");
+      });
+    });
+
+    describe("exclusion", () => {
+      class NoAdmin extends Model {
+        static {
+          this.attribute("role", "string");
+          this.validates("role", { exclusion: { in: ["admin", "root"] } });
+        }
+      }
+
+      it("accepts non-excluded values", async () => {
+        expect(await new NoAdmin({ role: "user" }).isValid()).toBe(true);
+      });
+
+      it("validates exclusion of", async () => {
+        const n = new NoAdmin({ role: "admin" });
+        expect(await n.isValid()).toBe(false);
+        expect(n.errors.get("role")).toContain("is reserved");
+      });
+    });
+
+    // -- Format --
+    describe("format", () => {
+      class Email extends Model {
+        static {
+          this.attribute("email", "string");
+          this.validates("email", { format: { with: /^[^@]+@[^@]+$/, multiline: true } });
+        }
+      }
+
+      it("validate format", async () => {
+        expect(await new Email({ email: "dean@example.com" }).isValid()).toBe(true);
+      });
+
+      it("rejects non-matching format", async () => {
+        const e = new Email({ email: "not-an-email" });
+        expect(await e.isValid()).toBe(false);
+        expect(e.errors.get("email")).toContain("is invalid");
+      });
+
+      it("skips null", async () => {
+        // Rails format validator does NOT auto-skip nil — opt in via
+        // allow_nil: true (matches EachValidator's allow_nil dispatch).
+        class NilSkippingEmail extends Model {
+          static {
+            this.attribute("email", "string");
+            this.validates("email", {
+              format: { with: /^[^@]+@[^@]+$/, multiline: true, allowNil: true },
+            });
+          }
+        }
+        expect(await new NilSkippingEmail({}).isValid()).toBe(true);
+      });
+    });
+
+    // -- Acceptance --
+    describe("acceptance", () => {
+      // Rails' equivalent uses a virtual attribute; boolean here so `"1"` /
+      // `true` round-trip through cast as `true` and match the default accept
+      // list `["1", true]`, whereas a string-typed attr would cast `true` to
+      // "t" (per type/immutable_string.rb) and rightly fail.
+      class Terms extends Model {
+        static {
+          this.attribute("accepted", "boolean");
+          this.validates("accepted", { acceptance: true });
+        }
+      }
+
+      it("terms of service agreement", async () => {
+        expect(await new Terms({ accepted: "1" }).isValid()).toBe(true);
+        expect(await new Terms({ accepted: true }).isValid()).toBe(true);
+      });
+
+      it("terms of service agreement no acceptance", async () => {
+        expect(await new Terms({ accepted: "0" }).isValid()).toBe(false);
+        expect(await new Terms({ accepted: false }).isValid()).toBe(false);
+      });
+
+      it("terms of service agreement with accept value", async () => {
+        class Custom extends Model {
+          static {
+            this.attribute("agreed", "string");
+            this.validates("agreed", {
+              acceptance: { accept: ["I agree", "yes"] },
+            });
+          }
+        }
+        expect(await new Custom({ agreed: "I agree" }).isValid()).toBe(true);
+        expect(await new Custom({ agreed: "no" }).isValid()).toBe(false);
+      });
+    });
+
+    // -- Confirmation --
+    describe("confirmation", () => {
+      class WithConfirm extends Model {
+        static {
+          this.attribute("password", "string");
+          this.validates("password", { confirmation: true });
+        }
+      }
+
+      it("passes when no confirmation field set", async () => {
+        expect(await new WithConfirm({ password: "secret" }).isValid()).toBe(true);
+      });
+
+      it("title confirmation", async () => {
+        expect(
+          await new WithConfirm({
+            password: "secret",
+            passwordConfirmation: "secret",
+          }).isValid(),
+        ).toBe(true);
+      });
+
+      it("no title confirmation", async () => {
+        const w = new WithConfirm({
+          password: "secret",
+          passwordConfirmation: "wrong",
+        });
+        expect(await w.isValid()).toBe(false);
+        expect(w.errors.get("passwordConfirmation")).toContain("doesn't match Password");
+      });
+    });
+
+    // -- Conditional validation --
+    describe("conditional", () => {
+      it("if validation using block false", async () => {
+        class Cond extends Model {
+          static {
+            this.attribute("name", "string");
+            this.attribute("requireName", "boolean", { default: false });
+            this.validates("name", {
+              presence: {
+                if: (r: any) => r.readAttribute("requireName") === true,
+              },
+            });
+          }
+        }
+        expect(await new Cond({ requireName: false }).isValid()).toBe(true);
+        expect(await new Cond({ requireName: true }).isValid()).toBe(false);
+      });
+
+      it("unless validation using block true", async () => {
+        class Unless extends Model {
+          static {
+            this.attribute("name", "string");
+            this.attribute("optional", "boolean", { default: false });
+            this.validates("name", {
+              presence: {
+                unless: (r: any) => r.readAttribute("optional") === true,
+              },
+            });
+          }
+        }
+        expect(await new Unless({ optional: true }).isValid()).toBe(true);
+        expect(await new Unless({ optional: false }).isValid()).toBe(false);
+      });
+    });
+
+    // -- Custom validate --
+    describe("custom validate", () => {
+      it("function validator", async () => {
+        class Custom extends Model {
+          static {
+            this.attribute("value", "integer");
+            this.validate(function (record: any) {
+              const val = record.readAttribute("value");
+              if (val !== null && (val as number) % 2 !== 0) {
+                record.errors.add("value", ":even", { message: "must be even" });
+              }
+            });
+          }
+        }
+        expect(await new Custom({ value: 4 }).isValid()).toBe(true);
+        const c = new Custom({ value: 3 });
+        expect(await c.isValid()).toBe(false);
+        expect(c.errors.get("value")).toContain("must be even");
+      });
+    });
+
+    // -- isInvalid --
+    it("invalid should be the opposite of valid", async () => {
+      class Required extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      expect(await new Required().isInvalid()).toBe(true);
+      expect(await new Required({ name: "dean" }).isInvalid()).toBe(false);
+    });
+
+    // -- fullMessages --
+    it("fullMessages prefixes attribute name", async () => {
+      class FM extends Model {
+        static {
+          this.attribute("title", "string");
+          this.validates("title", { presence: true });
+        }
+      }
+      const f = new FM();
+      await f.isValid();
+      expect(f.errors.fullMessages).toContain("Title can't be blank");
+    });
+
+    it("fullMessages for :base has no prefix", async () => {
+      class Base extends Model {
+        static {
+          this.validate((record: any) => {
+            record.errors.add("base", ":invalid", { message: "is broken" });
+          });
+        }
+      }
+      const b = new Base();
+      await b.isValid();
+      expect(b.errors.fullMessages).toContain("is broken");
+    });
+
+    // -- errors.clear between validations --
+    it("errors are cleared between isValid calls", async () => {
+      class Clearable extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      const c = new Clearable();
+      await c.isValid();
+      expect(c.errors.count).toBeGreaterThan(0);
+      c.writeAttribute("name", "dean");
+      await c.isValid();
+      expect(c.errors.count).toBe(0);
+    });
+  });
+  describe("custom messages", () => {
+    it("presence with custom message", async () => {
+      class Custom extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: { message: "is required" } });
+        }
+      }
+      const c = new Custom();
+      await c.isValid();
+      expect(c.errors.get("name")).toContain("is required");
+    });
+
+    it("length with custom tooShort and tooLong", async () => {
+      class Custom extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", {
+            length: { minimum: 3, maximum: 5, tooShort: "too few!", tooLong: "too many!" },
+          });
+        }
+      }
+      const short = new Custom({ name: "ab" });
+      await short.isValid();
+      expect(short.errors.get("name")).toContain("too few!");
+
+      const long = new Custom({ name: "abcdef" });
+      await long.isValid();
+      expect(long.errors.get("name")).toContain("too many!");
+    });
+  });
+  describe("errors.fullMessagesFor()", () => {
+    it("full_messages_for contains all the error messages for the given attribute indifferent", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("email", "string");
+          this.validates("name", { presence: true });
+          this.validates("email", { presence: true });
+        }
+      }
+      const u = new User({});
+      await u.isValid();
+      expect(u.errors.fullMessagesFor("name")).toEqual(["Name can't be blank"]);
+      expect(u.errors.fullMessagesFor("email")).toEqual(["Email can't be blank"]);
+      expect(u.errors.fullMessagesFor("other")).toEqual([]);
+    });
+  });
+
+  describe("errors.ofKind()", () => {
+    it("of_kind? defaults message to :invalid", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      const u = new User({});
+      await u.isValid();
+      expect(u.errors.ofKind("name", ":blank")).toBe(true);
+      expect(u.errors.ofKind("name", ":invalid")).toBe(false);
+      expect(u.errors.ofKind("name")).toBe(false);
+    });
+  });
+
+  describe("validators / validatorsOn", () => {
+    it("returns all registered validators", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("email", "string");
+          this.validates("name", { presence: true });
+          this.validates("email", { presence: true, length: { minimum: 5 } });
+        }
+      }
+      const validators = User.validators();
+      // presence on name, presence on email, length on email
+      expect(validators.length).toBe(3);
+    });
+
+    it("returns validators for a specific attribute", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("email", "string");
+          this.validates("name", { presence: true, length: { minimum: 2, maximum: 50 } });
+          this.validates("email", { presence: true });
+        }
+      }
+      const nameValidators = User.validatorsOn("name");
+      expect(nameValidators.length).toBe(2); // presence + length
+      const emailValidators = User.validatorsOn("email");
+      expect(emailValidators.length).toBe(1); // presence only
+    });
+
+    it("returns empty array for attribute with no validators", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("bio", "string");
+          this.validates("name", { presence: true });
+        }
+      }
+      expect(User.validatorsOn("bio")).toEqual([]);
+    });
+  });
+
+  describe("custom validation contexts", () => {
+    it("with a class that adds errors on create and validating a new model", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("terms_accepted", "string");
+          this.validates("name", { presence: true });
+          this.validates("terms_accepted", { presence: true, on: "registration" });
+        }
+      }
+      const u = new User({ name: "Alice" });
+      // Without context, terms_accepted validation is skipped
+      expect(await u.isValid()).toBe(true);
+      // With custom context, terms_accepted presence validation runs
+      expect(await u.isValid("registration")).toBe(false);
+    });
+
+    it("with a class that adds errors on update and validating a new model", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("email", "string");
+          this.validates("name", { presence: true });
+          this.validates("email", { presence: true, on: "create" });
+        }
+      }
+      const u = new User({ name: "Alice" });
+      expect(await u.isValid("create")).toBe(false);
+      expect(await u.isValid("update")).toBe(true);
+    });
+  });
+
+  describe("Errors enhancements", () => {
+    it("delete removes details on given attribute", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+        }
+      }
+      const u = new User({});
+      u.errors.add("name", ":blank");
+      u.errors.add("name", ":too_short");
+      u.errors.add("email", ":blank");
+      const removed = u.errors.delete("name");
+      expect(removed!.length).toBe(2);
+      expect(u.errors.count).toBe(1);
+    });
+
+    it("delete with type only removes matching errors", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+        }
+      }
+      const u = new User({});
+      u.errors.add("name", ":blank");
+      u.errors.add("name", ":too_short");
+      const removed = u.errors.delete("name", ":blank");
+      expect(removed!.length).toBe(1);
+      expect(u.errors.count).toBe(1);
+    });
+
+    it("each iterates over all errors", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+        }
+      }
+      const u = new User({});
+      u.errors.add("name", ":blank");
+      u.errors.add("email", ":invalid");
+      const collected: string[] = [];
+      u.errors.each((e) => collected.push(`${e.attribute}:${e.type}`));
+      expect(collected).toEqual(["name::blank", "email::invalid"]);
+    });
+
+    it("merge errors", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+        }
+      }
+      const u1 = new User({});
+      const u2 = new User({});
+      u1.errors.add("name", ":blank");
+      u2.errors.merge(u1.errors);
+      expect(u2.errors.count).toBe(1);
+      expect(u2.errors.get("name")).toEqual(["can't be blank"]);
+    });
+
+    it("to_hash returns the error messages hash", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("email", "string");
+        }
+      }
+      const u = new User({});
+      u.errors.add("name", ":blank");
+      u.errors.add("name", ":too_short");
+      u.errors.add("email", ":invalid");
+      const hash = u.errors.toHash();
+      expect(hash.get("name")!.length).toBe(2);
+      expect(hash.get("email")!.length).toBe(1);
+    });
+
+    it("include?", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+        }
+      }
+      const u = new User({});
+      u.errors.add("name", ":blank");
+      expect(u.errors.include("name")).toBe(true);
+      expect(u.errors.include("email")).toBe(false);
+    });
+
+    it("messages returns grouped messages", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+        }
+      }
+      const u = new User({});
+      u.errors.add("name", ":blank");
+      expect(u.errors.messages.get("name")).toEqual(["can't be blank"]);
+    });
+
+    it("full_messages creates a list of error messages with the attribute name included", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+        }
+      }
+      const u = new User({});
+      expect(u.errors.fullMessage("name", "is required")).toBe("Name is required");
+      expect(u.errors.fullMessage("base", "Something went wrong")).toBe("Something went wrong");
+    });
+  });
+
+  describe("conditional validates (if/unless)", () => {
+    it("skips validation when if condition returns false", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("requires_name", "boolean");
+          this.validates("name", {
+            presence: true,
+            if: (record: any) => record.readAttribute("requires_name") === true,
+          });
+        }
+      }
+      const u = new User({ requires_name: false });
+      expect(await u.isValid()).toBe(true);
+    });
+
+    it("runs validation when if condition returns true", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("requires_name", "boolean");
+          this.validates("name", {
+            presence: true,
+            if: (record: any) => record.readAttribute("requires_name") === true,
+          });
+        }
+      }
+      const u = new User({ requires_name: true });
+      expect(await u.isValid()).toBe(false);
+    });
+
+    it("skips validation when unless condition returns true", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("skip_validation", "boolean");
+          this.validates("name", {
+            presence: true,
+            unless: (record: any) => record.readAttribute("skip_validation") === true,
+          });
+        }
+      }
+      const u = new User({ skip_validation: true });
+      expect(await u.isValid()).toBe(true);
+    });
+
+    it("runs validation when unless condition returns false", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("skip_validation", "boolean");
+          this.validates("name", {
+            presence: true,
+            unless: (record: any) => record.readAttribute("skip_validation") === true,
+          });
+        }
+      }
+      const u = new User({ skip_validation: false });
+      expect(await u.isValid()).toBe(false);
+    });
+  });
+
+  describe("validates_*_of shorthand methods", () => {
+    it("validate presences", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("email", "string");
+          this.validatesPresenceOf("name", "email");
+        }
+      }
+      const u = new User({});
+      expect(await u.isValid()).toBe(false);
+      expect(u.errors.get("name").length).toBeGreaterThan(0);
+      expect(u.errors.get("email").length).toBeGreaterThan(0);
+    });
+
+    it("validates absence of", async () => {
+      class User extends Model {
+        static {
+          this.attribute("spam", "string");
+          this.validatesAbsenceOf("spam");
+        }
+      }
+      const u = new User({ spam: "not empty" });
+      expect(await u.isValid()).toBe(false);
+    });
+
+    it("validatesLengthOf validates length", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validatesLengthOf("name", { minimum: 3 });
+        }
+      }
+      expect(await new User({ name: "AB" }).isValid()).toBe(false);
+      expect(await new User({ name: "ABC" }).isValid()).toBe(true);
+    });
+
+    it("validatesSizeOf is an alias for validatesLengthOf", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validatesSizeOf("name", { minimum: 3 });
+        }
+      }
+      expect(await new User({ name: "AB" }).isValid()).toBe(false);
+      expect(await new User({ name: "ABC" }).isValid()).toBe(true);
+    });
+
+    it("validatesNumericalityOf validates numericality", async () => {
+      class Item extends Model {
+        static {
+          this.attribute("price", "float");
+          this.validatesNumericalityOf("price", { greaterThan: 0 });
+        }
+      }
+      expect(await new Item({ price: -1 }).isValid()).toBe(false);
+      expect(await new Item({ price: 10 }).isValid()).toBe(true);
+    });
+
+    it("validatesInclusionOf validates inclusion", async () => {
+      class User extends Model {
+        static {
+          this.attribute("role", "string");
+          this.validatesInclusionOf("role", { in: ["admin", "user"] });
+        }
+      }
+      expect(await new User({ role: "hacker" }).isValid()).toBe(false);
+      expect(await new User({ role: "admin" }).isValid()).toBe(true);
+    });
+
+    it("validatesFormatOf validates format", async () => {
+      class User extends Model {
+        static {
+          this.attribute("email", "string");
+          this.validatesFormatOf("email", { with: /@/ });
+        }
+      }
+      expect(await new User({ email: "nope" }).isValid()).toBe(false);
+      expect(await new User({ email: "a@b.com" }).isValid()).toBe(true);
+    });
+
+    it("validatesConfirmationOf validates confirmation", async () => {
+      class User extends Model {
+        static {
+          this.attribute("password", "string");
+          this.validatesConfirmationOf("password");
+        }
+      }
+      const u = new User({ password: "secret", passwordConfirmation: "mismatch" });
+      expect(await u.isValid()).toBe(false);
+    });
+
+    it("validatesLengthOf accepts multiple attributes", async () => {
+      class Topic extends Model {
+        static {
+          this.attribute("title", "string");
+          this.attribute("content", "string");
+          this.validatesLengthOf("title", "content", { minimum: 2 });
+        }
+      }
+      const t = new Topic({ title: "", content: "" });
+      expect(await t.isValid()).toBe(false);
+      expect(t.errors.get("title").length).toBeGreaterThan(0);
+      expect(t.errors.get("content").length).toBeGreaterThan(0);
+      expect(await new Topic({ title: "ok", content: "ok" }).isValid()).toBe(true);
+    });
+
+    it("validatesPresenceOf passes through strict option", async () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.validatesPresenceOf("name", { strict: true });
+        }
+      }
+      await expect(new User({}).isValid()).rejects.toThrow();
+      expect(await new User({ name: "Alice" }).isValid()).toBe(true);
+    });
+
+    it("validatesPresenceOf passes through if and on options", async () => {
+      class Topic extends Model {
+        static {
+          this.attribute("title", "string");
+          this.validatesPresenceOf("title", { if: () => true, on: "update" });
+        }
+      }
+      expect(await new Topic({ title: "" }).isValid()).toBe(true); // no context
+      expect(await new Topic({ title: "" }).isInvalid("update")).toBe(true);
+    });
+
+    it("validatesPresenceOf passes through unless option", async () => {
+      class Topic extends Model {
+        static {
+          this.attribute("title", "string");
+          this.validatesPresenceOf("title", { unless: () => true });
+        }
+      }
+      expect(await new Topic({ title: "" }).isValid()).toBe(true);
+    });
+  });
+
+  describe("Errors#generateMessage", () => {
+    it("generate_message works without i18n_scope", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+        }
+      }
+      const u = new User({});
+      expect(u.errors.generateMessage("name", ":blank")).toBe("can't be blank");
+      expect(u.errors.generateMessage("name", ":invalid")).toBe("is invalid");
+    });
+
+    it("substitutes options into message", () => {
+      class User extends Model {
+        static {
+          this.attribute("name", "string");
+          this.attribute("age", "integer");
+        }
+      }
+      const u = new User({});
+      expect(u.errors.generateMessage("age", ":greater_than", { count: 0 })).toBe(
+        "must be greater than 0",
+      );
+    });
+  });
+});
